requirements.txt
discord.py
lxml
urllib3